Renamed solution and project from DiscordApis to DiscordDashApis.  Added models to mirror the bot project.  Added aws tooling.
