+++ DiscordDashApis/Models/GuildUserActionEnum.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace DiscordBotApis.Models
+namespace DiscordDashApis.Models
 {
     public enum GuildUserActionEnum
     {
