+++ DiscordDashApis/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace DiscordBotApis
+namespace DiscordDashApis
 {
     public class Program
     {
