+++ DiscordDashApis/Startup.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using DiscordBotApis.Data;
+using DiscordDashApis.Data;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpsPolicy;
@@ -12,10 +12,10 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.EntityFrameworkCore;
-using DiscordBotApis.Data.Users;
+using DiscordDashApis.Data.Users;
 using Pomelo.EntityFrameworkCore;
 
-namespace DiscordBotApis
+namespace DiscordDashApis
 {
     public class Startup
     {
@@ -45,7 +45,7 @@ namespace DiscordBotApis
             services.AddScoped<IUserRepository, MockUserRepository>();
 
             services.AddDbContextPool<DiscordBotApiDbContext>(
-                options => options.UseMySql(Configuration["ConnectionString:DiscordBotData"])
+                options => options.UseSqlServer(Configuration["ConnectionString:DiscordBotData"])
             );
         }
