+++ DiscordDashApis/appsettings.json
@@ -5,7 +5,7 @@
     }
   },
   "ConnectionString": {
-    "DiscordBotData": "Server = (localdb)\\mssqllocaldb; Database = DiscordBotData; Trusted_Connection = True;"
+    "DiscordBotData": "connection string goes here"
   },
   "AllowedHosts": "*"
 }
