+++ DiscordDashApis/aws-beanstalk-tools-defaults.json
@@ -1,0 +1,19 @@
+
+{
+    "comment" : "This file is used to help set default values when using the dotnet CLI extension Amazon.ElasticBeanstalk.Tools. For more information run \"dotnet eb --help\" from the project root.",
+    "profile" : "default",
+    "region"  : "us-west-1",
+    "application" : "DiscordDashApis",
+    "environment" : "DiscordDashApis-dev",
+    "cname"       : "discorddashapis-dev2",
+    "solution-stack" : "64bit Windows Server 2016 v2.2.2 running IIS 10.0",
+    "environment-type" : "SingleInstance",
+    "instance-profile" : "aws-elasticbeanstalk-ec2-role",
+    "service-role"     : "aws-elasticbeanstalk-service-role",
+    "health-check-url" : "/",
+    "instance-type"    : "t3a.micro",
+    "key-pair"         : "discord-apis-key-pair",
+    "iis-website"      : "Default Web Site",
+    "app-path"         : "/",
+    "enable-xray"      : false
+}
